Finished Imput_By_Brand_and_Model
Tested the Imput_By_Brand_and_Model transformer and found a bug when calculating the mode for each group. If the mode was empty, the code would fail. Added a condition to check whether a mode exists; if not, NaN is returned in the mode_map.
Applied the same prevention for the original transformer.
Finished the doc string.